pyproject.toml
[project]
name = "hznu-os-ex4-nn"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "rich>=15.0.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = ["ruff>=0.15.10", "ty>=0.0.29"]

# formatter + linter
[tool.ruff]
line-length = 80
target-version = 'py312'

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "SIM", "TID", "PL", "RUF"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "lf"
